Moved LeaderboardEntry from nested in LeaderboardController to PetGame.Core Models
